test only in chrome
the new transactions test fail when ran in parallel, even when mode is set to serial, the same test runs in paralled across browsers which causes tests to fail.

an alternative solution would be to have a function that creates unique users for each test but for now the current setup is good enough.